docs(readme): update todo list
There is a general problem in which sometimes servers take a while to respond
to our request and may not be indicitive of the users REAL ping to the server.
It does not happen too often, but we need a way to rectify that if it does happen.
Might mark that for later work.